move shared Characteristic types
